+++ packages/coding-agent/test/suite/regressions/5943-session-start-notify.test.ts
@@ -17,6 +17,7 @@ function createUiContext(
 		notify: onNotify,
 		requestRender: () => {},
 		onTerminalInput: () => () => {},
+		onTranscriptReset: () => () => {},
 		setStatus: () => {},
 		setWorkingMessage: () => {},
 		setWorkingVisible: () => {},